07_코드/requirements.txt
requests>=2.31.0
pyyaml>=6.0
pykrx>=1.0.45
anthropic>=0.40.0
python-dateutil>=2.8.2
dart-fss>=0.4.4
